// InvoicerClient/src/app/api/model/getAllPaymentsResponse.ts
/**
 * Invoicer API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface GetAllPaymentsResponse {
  id?: string;
  amount?: number;
  paidOn?: string;
  notes?: string | null;
  invoiceId?: string;
  invoiceNumber?: string | null;
  clientName?: string | null;
}